# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pay me Back",
        "description": "API",
        "contact": {
            "email": "flailton@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "API Local Server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Consultar usuários",
                "description": "Consultar todos os usuários",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "int",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "admin"
                                                    },
                                                    "email": {
                                                        "type": "email",
                                                        "example": "admin@admin.com"
                                                    },
                                                    "document": {
                                                        "type": "string",
                                                        "example": "991.290.566-73"
                                                    },
                                                    "user_type_id": {
                                                        "type": "int",
                                                        "example": "1"
                                                    },
                                                    "user_type": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "int",
                                                                "example": "1"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Usuário"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Criar usuário",
                "description": "Criar usuário na API",
                "operationId": "store",
                "requestBody": {
                    "description": "Criar usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "document",
                                    "user_type_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin@123"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "admin@admin.com"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "99129056673"
                                    },
                                    "user_type_id": {
                                        "type": "int",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "properties": {
                                                "id": {
                                                    "type": "int",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "admin"
                                                },
                                                "email": {
                                                    "type": "email",
                                                    "format": "email",
                                                    "example": "admin@admin.com"
                                                },
                                                "document": {
                                                    "type": "string",
                                                    "example": "991.290.566-73"
                                                },
                                                "user_type_id": {
                                                    "type": "int",
                                                    "example": "1"
                                                }
                                            },
                                            "type": "object",
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "O Token está expirado!"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados fora do padrão esperado ou não informados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo nome é obrigatório!"
                                                    },
                                                    "collectionFormat": "multi"
                                                }
                                            },
                                            "type": "object",
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}